+++ test/test_data/get_active_iac.json
@@ -1,0 +1,7 @@
+{
+  "iac": "jkbvyklkwj88",
+  "active": true,
+  "lastUsedDateTime": "2017-05-15T10:00:00Z",
+  "caseId": "7bc5d41b-0549-40b3-ba76-42f6d4cf3fdb",
+  "questionSet": "H1"
+}
