+++ test/test_data/get_updated_iac.json
@@ -1,0 +1,9 @@
+{
+    "code": "jkbvyklkwj88",
+    "active": false,
+    "createdBy": "SYSTEM",
+    "createdDateTime": 1534151837213,
+    "updatedBy": "SYSTEM",
+    "updatedDateTime": 1534168018020,
+    "lastUsedDateTime": null
+}
